requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ollama==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
